Rename WellSeparatedPairDecomposition to WSPD.
