+++ hgeometry/hgeometry.cabal
@@ -137,6 +137,7 @@ library
                     Algorithms.Geometry.EuclideanMST
                     Algorithms.Geometry.EuclideanMST.EuclideanMST
 
+                    Algorithms.Geometry.WSPD
                     Algorithms.Geometry.WellSeparatedPairDecomposition.WSPD
                     Algorithms.Geometry.WellSeparatedPairDecomposition.Types
 
@@ -189,6 +190,8 @@ library
 
                     Algorithms.Geometry.LineSegmentIntersection.Types
                     Algorithms.Geometry.SmallestEnclosingBall.Types
+
+                    Algorithms.Geometry.WSPD.Types
 
                     Data.Geometry.Point.Internal
                     Data.Geometry.Point.Orientation
