+++ hgeometry/src/Algorithms/Geometry/EuclideanMST/EuclideanMST.hs
@@ -11,7 +11,7 @@
 --------------------------------------------------------------------------------
 module Algorithms.Geometry.EuclideanMST.EuclideanMST
   {-# DEPRECATED "This module will be deleted after 2021-06-01. \
-                 \Use Algorithms.Geometry.EuclideanMST instead" #-}
+                 \Use Algorithms.Geometry.EuclideanMST instead." #-}
   ( module Algorithms.Geometry.EuclideanMST ) where
 
 import Algorithms.Geometry.EuclideanMST
